add migration for templates
